Rename Phase 29 shadow ML modules
