+++ control-plane/aegisops_control_plane/ml_shadow/__init__.py
@@ -1,1 +1,5 @@
-"""Package scaffold only for future reviewed ML shadow-mode modules."""
+"""Reviewed ML shadow-mode package.
+
+The root-level ``phase29_*`` modules remain compatibility shims for legacy
+callers; new internal imports should use the domain-named modules here.
+"""
