+++ control-plane/tests/test_cross_boundary_negative_e2e_validation.py
@@ -30,11 +30,11 @@ from aegisops_control_plane.models import (
     RecommendationRecord,
     ReconciliationRecord,
 )
-from aegisops_control_plane.phase29_evidently_drift_visibility import (
+from aegisops_control_plane.ml_shadow.drift_visibility import (
     build_phase29_evidently_drift_visibility_report,
 )
-from aegisops_control_plane.phase29_shadow_dataset import Phase29ShadowDatasetSnapshot
-from aegisops_control_plane.phase29_shadow_scoring import (
+from aegisops_control_plane.ml_shadow.dataset import Phase29ShadowDatasetSnapshot
+from aegisops_control_plane.ml_shadow.scoring import (
     Phase29ShadowScoringError,
     score_shadow_dataset_offline,
 )
