+++ control-plane/tests/test_phase29_mlflow_shadow_model_registry_validation.py
@@ -14,13 +14,13 @@ for candidate in (CONTROL_PLANE_ROOT, TESTS_ROOT):
         sys.path.insert(0, str(candidate))
 
 
-import aegisops_control_plane.phase29_mlflow_shadow_model_registry as mlflow_shadow_model_registry
+import aegisops_control_plane.ml_shadow.mlflow_registry as mlflow_shadow_model_registry
 from aegisops_control_plane.models import EvidenceRecord, ReconciliationRecord
-from aegisops_control_plane.phase29_mlflow_shadow_model_registry import (
+from aegisops_control_plane.ml_shadow.mlflow_registry import (
     Phase29MlflowShadowModelRegistryError,
     track_shadow_model_with_mlflow,
 )
-from aegisops_control_plane.phase29_shadow_dataset import (
+from aegisops_control_plane.ml_shadow.dataset import (
     Phase29ShadowDatasetSnapshot,
     generate_reviewed_shadow_dataset,
 )
